refactor: update protected route
